test: 删除零风险重复测试文件，精简测试库
- 删除 test_comprehensive_final.py（6 例全被覆盖的冒烟测试）
- 删除 test_database_extended.py（6 例废测试：不测 app 代码）
- 删除 test_sync_full.py（4 例被 test_sync_service* 覆盖）
- 删除 test_data_util_extended.py（1 例被 test_data_util.py 覆盖）
- test_sync_more.py 补充 CustomItem 默认值断言（source/ori_title=None）
